Copy prometheus test client from cluster-openshift-monitoring
This is intended to simplify e2e testing of operator configuration of
metrics endpoints and prometheus monitoring of those endpoints.

The Poll helper dependency was also copied since it seems generally
useful.

Use of pkg/errors wrapping was switched to golang >= 1.13's wrapped
errors to avoid adding a dependency that most operators are unlikely
to use.